improvements part editing
The biggest way this project can be improved is by mathematically solving for the data, which would not only be more accurate, but could also prove if there is causation with the trends. I used features selections models I searched online and compare the result with my insight and it show a interesting result.